2025/globant-gdg/image-generation/requirements.txt
streamlit==1.24.0
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0